requirements.txt
mysql-connector-python
pwinput
python-dotenv
tabulate